Refactor redirects and auth to middlewares
