+++ httpx/client.py
@@ -43,6 +43,12 @@ from .models import (
     URLTypes,
 )
 from .status_codes import codes
+from .middlewares import (
+    BaseMiddleware,
+    RedirectMiddleware,
+    HTTPBasicAuthMiddleware,
+    HTTPDigestAuthMiddleware,
+)
 
 
 class BaseClient:
@@ -61,6 +67,7 @@ class BaseClient:
         app: typing.Callable = None,
         raise_app_exceptions: bool = True,
         backend: ConcurrencyBackend = None,
+        additional_middleware: list = None,
     ):
         if backend is None:
             backend = AsyncioBackend()
@@ -131,32 +138,43 @@ class BaseClient:
         cert: CertTypes = None,
         timeout: TimeoutTypes = None,
     ) -> AsyncResponse:
-        if auth is None:
-            auth = self.auth
-
         url = request.url
 
         if url.scheme not in ("http", "https"):
             raise InvalidURL('URL scheme must be "http" or "https".')
 
+        middlewares: typing.List[BaseMiddleware] = [
+            RedirectMiddleware(
+                allow_redirects=allow_redirects,
+                max_redirects=self.max_redirects,
+                base_cookies=self.cookies,
+            ),
+        ]
         if auth is None and (url.username or url.password):
-            auth = HTTPBasicAuth(username=url.username, password=url.password)
+            middlewares.append(HTTPBasicAuthMiddleware(username=url.username, password=url.password))
+        elif isinstance(auth, tuple):
+            middlewares.append(HTTPBasicAuthMiddleware(username=auth[0], password=auth[1]))
+        elif isinstance(auth, HTTPBasicAuth):
+            middlewares.append(HTTPBasicAuthMiddleware(username=auth.username, password=auth.password))
+        elif isinstance(auth, HTTPDigestAuth):
+            middlewares.append(HTTPDigestAuthMiddleware(username=auth.username, password=auth.password))
 
-        if auth is not None:
-            if isinstance(auth, HTTPDigestAuth):
-                self.auth = auth
-            else:
-                if isinstance(auth, tuple):
-                    auth = HTTPBasicAuth(username=auth[0], password=auth[1])
-                request = auth(request)
+        while True:
+            for middleware in middlewares:
+                request = middleware.process_request(request)
 
-        response = await self.send_handling_redirects(
-            request,
-            verify=verify,
-            cert=cert,
-            timeout=timeout,
-            allow_redirects=allow_redirects,
-        )
+            response = await self.dispatch.send(
+                request, verify=verify, cert=cert, timeout=timeout
+            )
+
+            for middleware in middlewares:
+                result = middleware.process_response(request, response)
+                if isinstance(result, AsyncRequest):
+                    request = result
+                    break  # do not process more middlewares as one has an action to take
+
+            if isinstance(result, AsyncResponse):
+                break  # we have a response after all middlewares have processed the result
 
         if not stream:
             try:
@@ -165,147 +183,6 @@ class BaseClient:
                 await response.close()
 
         return response
-
-    async def send_handling_redirects(
-        self,
-        request: AsyncRequest,
-        *,
-        cert: CertTypes = None,
-        verify: VerifyTypes = None,
-        timeout: TimeoutTypes = None,
-        allow_redirects: bool = True,
-        history: typing.List[AsyncResponse] = None,
-    ) -> AsyncResponse:
-        if history is None:
-            history = []
-
-        while True:
-            # We perform these checks here, so that calls to `response.next()`
-            # will raise redirect errors if appropriate.
-            if len(history) > self.max_redirects:
-                raise TooManyRedirects()
-            if request.url in [response.url for response in history]:
-                raise RedirectLoop()
-
-            response = await self.dispatch.send(
-                request, verify=verify, cert=cert, timeout=timeout
-            )
-            # TODO: requests also checks for number of 401 calls < 2
-            if response.is_client_error and response.expects_digest_auth:
-                request = self.auth(request, response)  # TODO: remove dependency on self.auth
-                response = await self.dispatch.send(
-                    request, verify=verify, cert=cert, timeout=timeout
-                )
-
-            should_close_response = True
-            try:
-                assert isinstance(response, AsyncResponse)
-                response.history = list(history)
-                self.cookies.extract_cookies(response)
-                history = history + [response]
-
-                if allow_redirects and response.is_redirect:
-                    request = self.build_redirect_request(request, response)
-                else:
-                    should_close_response = False
-                    break
-            finally:
-                if should_close_response:
-                    await response.close()
-
-        if response.is_redirect:
-
-            async def send_next() -> AsyncResponse:
-                nonlocal request, response, verify, cert
-                nonlocal allow_redirects, timeout, history
-                request = self.build_redirect_request(request, response)
-                response = await self.send_handling_redirects(
-                    request,
-                    allow_redirects=allow_redirects,
-                    verify=verify,
-                    cert=cert,
-                    timeout=timeout,
-                    history=history,
-                )
-                return response
-
-            response.next = send_next  # type: ignore
-
-        return response
-
-    def build_redirect_request(
-        self, request: AsyncRequest, response: AsyncResponse
-    ) -> AsyncRequest:
-        method = self.redirect_method(request, response)
-        url = self.redirect_url(request, response)
-        headers = self.redirect_headers(request, url)
-        content = self.redirect_content(request, method)
-        cookies = self.merge_cookies(request.cookies)
-        return AsyncRequest(
-            method=method, url=url, headers=headers, data=content, cookies=cookies
-        )
-
-    def redirect_method(self, request: AsyncRequest, response: AsyncResponse) -> str:
-        """
-        When being redirected we may want to change the method of the request
-        based on certain specs or browser behavior.
-        """
-        method = request.method
-
-        # https://tools.ietf.org/html/rfc7231#section-6.4.4
-        if response.status_code == codes.SEE_OTHER and method != "HEAD":
-            method = "GET"
-
-        # Do what the browsers do, despite standards...
-        # Turn 302s into GETs.
-        if response.status_code == codes.FOUND and method != "HEAD":
-            method = "GET"
-
-        # If a POST is responded to with a 301, turn it into a GET.
-        # This bizarre behaviour is explained in 'requests' issue 1704.
-        if response.status_code == codes.MOVED_PERMANENTLY and method == "POST":
-            method = "GET"
-
-        return method
-
-    def redirect_url(self, request: AsyncRequest, response: AsyncResponse) -> URL:
-        """
-        Return the URL for the redirect to follow.
-        """
-        location = response.headers["Location"]
-
-        url = URL(location, allow_relative=True)
-
-        # Facilitate relative 'Location' headers, as allowed by RFC 7231.
-        # (e.g. '/path/to/resource' instead of 'http://domain.tld/path/to/resource')
-        if url.is_relative_url:
-            url = request.url.join(url)
-
-        # Attach previous fragment if needed (RFC 7231 7.1.2)
-        if request.url.fragment and not url.fragment:
-            url = url.copy_with(fragment=request.url.fragment)
-
-        return url
-
-    def redirect_headers(self, request: AsyncRequest, url: URL) -> Headers:
-        """
-        Strip Authorization headers when responses are redirected away from
-        the origin.
-        """
-        headers = Headers(request.headers)
-        if url.origin != request.url.origin:
-            del headers["Authorization"]
-        return headers
-
-    def redirect_content(self, request: AsyncRequest, method: str) -> bytes:
-        """
-        Return the body that should be used for the redirect request.
-        """
-        if method != request.method and method == "GET":
-            return b""
-        if request.is_streaming:
-            raise RedirectBodyUnavailable()
-        return request.content
 
 
 class AsyncClient(BaseClient):
