+++ httpx/middlewares/__init__.py
@@ -1,0 +1,10 @@
+from .base import BaseMiddleware
+from .redirect import RedirectMiddleware
+from .auth import HTTPBasicAuthMiddleware, HTTPDigestAuthMiddleware
+
+__all__ = [
+    "BaseMiddleware",
+    "RedirectMiddleware",
+    "HTTPBasicAuthMiddleware",
+    "HTTPDigestAuthMiddleware",
+]
